requirements.txt
functions-framework==3.3.0
firebase-admin==6.2.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.11.1
google-cloud-pubsub==2.17.1
python-dotenv==1.0.0
aiohttp==3.8.5
async-timeout==4.0.3
requests==2.31.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0